Store one canonical unit, and convert only where she types (#82)
Metric or imperial from the device locale, overridable in Profile ▸ Units,
with feet-and-inches and stones-and-pounds entered as two fields rather than
as a decimal box.

The rule everything else here serves: **every body measurement is stored in
SI** — `weightKg` in kilograms, `heightCm` in centimeters — and the setting
decides only what a screen draws. If a stored value depended on the
preference, switching it would rewrite history, and the rewrite would be
invisible, because the wrong number is a plausible number: a weight logged as
150 lb coming back as 150 kg. `EvaBodyUnits` is the whole boundary, two
functions per quantity, one each way, and `EvaBodyInputTests` encodes the
request body before and after a tour of all three systems and compares the
bytes.

Two consequences that look like details and are not.

The canonical value is a `Double`. A kilogram is 2.2 lb, so at whole
kilograms 150 lb and 151 lb are the same stored number and both come back as
150 — losing a pound silently, which is the same defect as 5'9" becoming 5.9.
Stored values land on a 0.01 kg / 0.1 cm grid, fine enough that every pound
and every inch round-trips exactly. The wire format does not change:
`parseProfile` has always taken a finite number in a range rather than an
integer, and `JSONEncoder` writes a whole `Double` as `64`, so a metric
profile sends the bytes it always sent and no API change was needed.

And a compound unit is two rows. Feet steps the height by twelve inches,
inches by one, both against the same total, so 5'11" + 1 in is 6'0".

Three systems where the canvas' Settings row draws two, because the PRD asks
for pounds and #82's decision asks for stones as well, and one word cannot
mean both without the setting depending on the locale again. `en_GB` defaults
to metric, which is the issue's own acceptance criterion and not what
`Locale.MeasurementSystem.uk` would suggest.

`EvaRadioRow` is DESIGN.md §6's radio row, built for the option list and
added to the specimen. The preference lives in `UserDefaults`, not on
`users/{uid}`: an account field is an always-human schema gate and #82 put
any change to stored data out of scope. ARCHITECTURE §5 states the canonical
unit and says what moving it onto the account would cost.

One shared test helper changed. `scrollIntoView` stopped as soon as
`isHittable` was true, which it is for an element merely overhanging the
bottom of the window — the tap then lands on the tab bar and the test fails
several assertions later. The settings row pushed `destructive.Delete
profile` to y 827…879 in an 874-point window and `DeleteAccountUITests`
failed 3 runs in 3 on it. The loop now also runs while the element's `maxY`
is past the window's, which only ever adds a scroll.

+++ mobile/Eva/Networking/APIModels.swift
@@ -68,8 +68,13 @@ struct APIUser: Codable {
 /// API uses it to resolve which day "today" is when it checks the 18+ floor.
 struct ProfilePayload: Codable {
     let dateOfBirth: String
-    let weightKg: Int
-    let heightCm: Int
+    /// SI, always — the device converts at the edge and never stores what it displayed
+    /// (#82, `EvaBodyUnits`). `Double` because whole kilograms cannot represent a pound:
+    /// `parseProfile` has always validated these as finite numbers in a range rather than
+    /// as integers, and `JSONEncoder` writes a whole `Double` as `64`, so a metric entry
+    /// sends the bytes it always sent.
+    let weightKg: Double
+    let heightCm: Double
     let goals: [String]
     let conditions: [String]
     let medications: String
